Add files via upload
For this We will implement the Deep Q-Learning algorithm with Experience Replay as described in breakthrough paper __"Playing Atari with Deep Reinforcement Learning"__. We will be testing the Q-Learning algorithm in the __"Highway Environment"__. In this environment our Q-Learning agent will operate a green car on a highway with other cars. The agent is penalized for hitting other cars and rewarded for driving faster.